Allow fractional delay for blocking mode changes and fix a few smaller memory leaks

+++ src/timers.h
@@ -17,8 +17,8 @@
 void timer_start(const enum timers i);
 double timer_elapsed_msec(const enum timers i);
 void sleepms(const int milliseconds);
-void set_blockingmode_timer(int delay, bool blocked);
-void get_blockingmode_timer(int *delay, bool *target_status);
+void set_blockingmode_timer(double delay, bool blocked);
+void get_blockingmode_timer(double *delay, bool *target_status);
 void *timer(void *val);
 unsigned long converttimeval(const struct timeval time) __attribute__((const));
